bugfix: Ignore workspace where no java project exists(#321)
